requirements.txt
pytest==8.2.0
requests==2.31.0
pydantic==2.7.4
allure-pytest==2.13.5
allure-python-commons==2.13.5